chore(refactoring): Move document code to runtime-core
