+++ connector-runtime/connector-runtime-core/pom.xml
@@ -121,6 +121,11 @@
       <artifactId>jsonassert</artifactId>
       <scope>test</scope>
     </dependency>
+    <dependency>
+      <groupId>org.wiremock</groupId>
+      <artifactId>wiremock-standalone</artifactId>
+      <scope>test</scope>
+    </dependency>
 
   </dependencies>
 </project>
