+++ connector-runtime/connector-runtime-core/src/main/java/io/camunda/connector/runtime/core/document/DocumentMetadataImpl.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.camunda.document;
+package io.camunda.connector.runtime.core.document;
 
 import io.camunda.connector.api.document.DocumentMetadata;
 import java.time.OffsetDateTime;
