+++ connector-runtime/connector-runtime-core/src/main/java/io/camunda/connector/runtime/core/document/store/InMemoryDocumentStore.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.camunda.document.store;
+package io.camunda.connector.runtime.core.document.store;
 
 import io.camunda.connector.api.document.DocumentCreationRequest;
 import io.camunda.connector.api.document.DocumentLinkParameters;
