+++ connector-runtime/connector-runtime-core/src/test/java/io/camunda/connector/runtime/core/document/DocumentHelperTest.java
@@ -14,9 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.camunda.document;
-
-import static org.mockito.Mockito.*;
+package io.camunda.connector.runtime.core.document;
 
 import io.camunda.connector.api.document.Document;
 import io.camunda.connector.api.document.DocumentCreationRequest;
