+++ connector-runtime/connector-runtime-core/src/test/java/io/camunda/connector/runtime/core/document/HttpClientServiceTest.java
